requirements.txt
langchain
google-generativeai
python-dotenv
sentence-transformers
streamlit
faiss-cpu
langchain_community
PyPDF2